# iterator: optimize merged iterator (again)