* specialize index heap with int type
* use heap fix instead of pop&push

benchmarks:
```bash
$ benchstat master.txt pr.txt 
name                       old time/op  new time/op  delta
MergedIterator/2_iters-8   8.97µs ± 0%  6.34µs ± 0%  -29.38%  (p=0.008 n=5+5)
MergedIterator/50_iters-8   732µs ± 0%   580µs ± 1%  -20.67%  (p=0.008 n=5+5)
```